requirement.txt
python-dotenv 
groq
streamlit